Add files via upload
Reupload original extra events with EventClass

Note the ideology events are all EVENT_CLASS_BAD due to the fact that some enable definitively bad events for other players, and so have to be classed as bad (otherwise one of the options would not fire, should bad events be disabled). Since some are bad, all have to be bad for ideology balance.